Try to fix single-threaded builds
