+++ crates/bevy_tasks/Cargo.toml
@@ -13,7 +13,9 @@ multi-threaded = ["dep:async-channel", "dep:async-task", "dep:concurrent-queue"]
 
 [dependencies]
 futures-lite = "2.0.1"
-async-executor = { git = "https://github.com/james7132/async-executor", branch = "leaked-executor", features = ["static"] }
+async-executor = { git = "https://github.com/james7132/async-executor", branch = "leaked-executor", features = [
+  "static",
+] }
 async-channel = { version = "2.2.0", optional = true }
 async-io = { version = "2.0.0", optional = true }
 async-task = { version = "4.2.0", optional = true }
