+++ crates/bevy_tasks/src/single_threaded_task_pool.rs
@@ -5,6 +5,14 @@ thread_local! {
     static LOCAL_EXECUTOR: async_executor::LocalExecutor<'static> = async_executor::LocalExecutor::new();
 }
 
+/// A [`TaskPool`] optimized for use in static variables.
+pub type StaticTaskPool = TaskPool;
+
+/// A [`StaticTaskPool`] scope for running one or more non-`'static` futures.
+///
+/// For more information, see [`TaskPool::scope`].
+pub type StaticScope<'scope, 'env, T> = Scope<'scope, 'env, T>;
+
 /// Used to create a [`TaskPool`].
 #[derive(Debug, Default, Clone)]
 pub struct TaskPoolBuilder {}
@@ -25,8 +33,8 @@ impl<'a> ThreadExecutor<'a> {
 
 impl TaskPoolBuilder {
     /// Creates a new `TaskPoolBuilder` instance
-    pub fn new() -> Self {
-        Self::default()
+    pub const fn new() -> Self {
+        Self {}
     }
 
     /// No op on the single threaded task pool
@@ -45,7 +53,7 @@ impl TaskPoolBuilder {
     }
 
     /// Creates a new [`TaskPool`]
-    pub fn build(self) -> TaskPool {
+    pub const fn build(self) -> TaskPool {
         TaskPool::new_internal()
     }
 }
@@ -62,13 +70,25 @@ impl TaskPool {
     }
 
     /// Create a `TaskPool` with the default configuration.
-    pub fn new() -> Self {
+    pub const fn new() -> Self {
         TaskPoolBuilder::new().build()
     }
 
     #[allow(unused_variables)]
-    fn new_internal() -> Self {
+    const fn new_internal() -> Self {
         Self {}
+    }
+
+    /// Checks if the threads in the task pool have been started or not. This always returns
+    /// true in single threaded builds.
+    pub fn is_initialized(&self) -> bool {
+        true
+    }
+
+    /// Initializes the task pool with the provided builder. This always is a no-op
+    /// true in single threaded builds.
+    #[allow(unused_variables)]
+    pub fn init(&self, builder: TaskPoolBuilder) {
     }
 
     /// Return the number of threads owned by the task pool
